refactor(ios): remove protocol javascriptexecutor
