+++ ios/sdk/base/executors/HippyJSCExecutor.h
@@ -21,11 +21,18 @@
  */
 
 #import <JavaScriptCore/JavaScriptCore.h>
-
-#import "HippyJavaScriptExecutor.h"
+#import "HippyDefines.h"
+#import "HippyBridgeModule.h"
+#import "HippyInvalidating.h"
+#import <memory>
 
 typedef void (^HippyJavaScriptValueCallback)(JSValue *result, NSError *error);
 
+typedef void (^HippyJavaScriptCompleteBlock)(NSError *error);
+typedef void (^HippyJavaScriptCallback)(id result, NSError *error);
+
+class Scope;
+@class HippyBridge;
 /**
  * Default name for the JS thread
  */
@@ -65,7 +72,20 @@ HIPPY_EXTERN NSString *const HippyFBJSValueClassKey;
 /**
  * Uses a JavaScriptCore context as the execution engine.
  */
-@interface HippyJSCExecutor : NSObject <HippyJavaScriptExecutor>
+@interface HippyJSCExecutor : NSObject<HippyBridgeModule, HippyInvalidating>
+
+@property (nonatomic, strong) HippyBridge *bridge;
+/**
+ * Whether the executor has been invalidated
+ */
+@property (nonatomic, readonly, getter=isValid) BOOL valid;
+
+@property (nonatomic, copy) NSString *executorkey;
+/*
+ *hippy-core js engine
+ */
+@property (atomic, assign) std::shared_ptr<Scope> pScope;
+@property (readonly) JSGlobalContextRef JSGlobalContextRef;
 
 /**
  * Returns whether executor uses custom JSC library.
@@ -80,6 +100,14 @@ HIPPY_EXTERN NSString *const HippyFBJSValueClassKey;
  */
 @property (nonatomic, copy) NSString *contextName;
 
+- (instancetype)initWithExecurotKey:(NSString *)execurotkey bridge:(HippyBridge *)bridge;
+
+/**
+ * Used to set up the executor after the bridge has been fully initialized.
+ * Do any expensive setup in this method instead of `-init`.
+ */
+- (void)setUp;
+
 /**
  * Invokes the given module/method directly. The completion block will be called with the
  * JSValue returned by the JS context.
@@ -92,5 +120,55 @@ HIPPY_EXTERN NSString *const HippyFBJSValueClassKey;
              jsValueCallback:(HippyJavaScriptValueCallback)onComplete;
 
 - (JSValueRef)JSTurboObjectWithName:(NSString *)name;
+
+/**
+ * Executes BatchedBridge.flushedQueue on JS thread and calls the given callback
+ * with JSValue, containing the next queue, and JSContext.
+ */
+- (void)flushedQueue:(HippyJavaScriptCallback)onComplete;
+
+/**
+ * called when second bundle load
+ */
+- (void)secondBundleLoadCompleted:(BOOL)success;
+
+/**
+ * called before excute secondary js bundle
+ */
+- (void)updateGlobalObjectBeforeExcuteSecondary;
+
+/**
+ * Executes BatchedBridge.callFunctionReturnFlushedQueue with the module name,
+ * method name and optional additional arguments on the JS thread and calls the
+ * given callback with JSValue, containing the next queue, and JSContext.
+ */
+- (void)callFunctionOnModule:(NSString *)module method:(NSString *)method arguments:(NSArray *)args callback:(HippyJavaScriptCallback)onComplete;
+
+/**
+ * Executes BatchedBridge.invokeCallbackAndReturnFlushedQueue with the cbID,
+ * and optional additional arguments on the JS thread and calls the
+ * given callback with JSValue, containing the next queue, and JSContext.
+ */
+- (void)invokeCallbackID:(NSNumber *)cbID arguments:(NSArray *)args callback:(HippyJavaScriptCallback)onComplete;
+
+/**
+ * Runs an application script, and notifies of the script load being complete via `onComplete`.
+ */
+- (void)executeApplicationScript:(NSData *)script sourceURL:(NSURL *)sourceURL onComplete:(HippyJavaScriptCompleteBlock)onComplete;
+
+- (void)injectJSONText:(NSString *)script asGlobalObjectNamed:(NSString *)objectName callback:(HippyJavaScriptCompleteBlock)onComplete;
+
+/**
+ * Enqueue a block to run in the executors JS thread. Fallback to `dispatch_async`
+ * on the main queue if the executor doesn't own a thread.
+ */
+- (void)executeBlockOnJavaScriptQueue:(dispatch_block_t)block;
+
+/**
+ * Special case for Timers + ContextExecutor - instead of the default
+ *   if jsthread then call else dispatch call on jsthread
+ * ensure the call is made async on the jsthread
+ */
+- (void)executeAsyncBlockOnJavaScriptQueue:(dispatch_block_t)block;
 
 @end
